ModelNameWithJavaKeywordTest,ImportedTypesWithNoFieldDefinitions, EntityWithSameQueryNameModel movet to SpecialCasesModel
