--- judo-runtime-core-jsl-itest/models/EntityWithSameQueryNameModel/src/test/resources/EntityWithSameQueryName.jsl
@@ -1,15 +1,0 @@
-model EntityWithSameQueryName;
-
-import judo::types;
-
-entity ReferenceEntity {
-    field String name;
-}
-entity entityA {
-    @Query
-    relation ReferenceEntity querySameName(String name) <= ReferenceEntity!any();
-}
-entity entityB {
-    @Query
-    relation ReferenceEntity querySameName(String name) <= ReferenceEntity!any();
-}
