--- judo-runtime-core-jsl-itest/models/ImportedTypesWithNoFieldDefinitions/src/test/java/logback.xml
@@ -1,34 +1,0 @@
-<!--
-  #%L
-  JUDO Runtime Core JSL :: Parent
-  %%
-  Copyright (C) 2018 - 2022 BlackBelt Technology
-  %%
-  This program and the accompanying materials are made available under the
-  terms of the Eclipse Public License 2.0 which is available at
-  http://www.eclipse.org/legal/epl-2.0.
-  
-  This Source Code may also be made available under the following Secondary
-  Licenses when the conditions for such availability set forth in the Eclipse
-  Public License, v. 2.0 are satisfied: GNU General Public License, version 2
-  with the GNU Classpath Exception which is
-  available at https://www.gnu.org/software/classpath/license.html.
-  
-  SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
-  #L%
-  -->
-<configuration>
-
-    <appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
-        <layout class="ch.qos.logback.classic.PatternLayout">
-            <Pattern>
-                %d{HH:mm:ss.SSS} [%t] %-5level %logger{36} - %msg%n
-            </Pattern>
-        </layout>
-    </appender>
-
-    <root level="error">
-        <appender-ref ref="CONSOLE"/>
-    </root>
-
-</configuration>
