+++ judo-runtime-core-jsl-itest/models/SpecialCasesModel/src/test/resources/SpecialCases.jsl
@@ -1,0 +1,31 @@
+model SpecialCases;
+
+import judo::types;
+
+// Imported types with no field definition
+
+entity TestEntity {
+    field Boolean dateSmaller = Date!of(year = 2023, month = 4, day = 17) < Date!of(year = 2023, month = 4, day = 19);
+    field Boolean timeSmaller = Time!of(hour = 11, minute = 11) < Time!of(hour = 11, minute = 19);
+    field Boolean timestampSmaller = Timestamp!of(date = `2023-04-17`) < Timestamp!of(date = `2023-04-19`);
+}
+
+// Model name with java keyword test
+
+entity Abstract {
+
+}
+
+// Entity with the same query name
+
+entity ReferenceEntity {
+    field String name;
+}
+entity entityA {
+    @Query
+    relation ReferenceEntity querySameName(String name) <= ReferenceEntity!any();
+}
+entity entityB {
+    @Query
+    relation ReferenceEntity querySameName(String name) <= ReferenceEntity!any();
+}
