+++ judo-runtime-core-jsl-itest/models/pom.xml
@@ -32,17 +32,17 @@
         <module>MappedTransferAssociationAssociationModel</module>
         <module>MappedTransferCompositonAggregationModel</module>
         <module>MappedTransferCompositonAssociationModel</module>
-        <module>MappedTransferOperatorAndFunctionOfEnumTestModel</module>
         <module>MappedTransferPrimitivesModel</module>
         <module>ModelNameWithJavaKeyword</module>
         <module>NavigationTestModel</module>
-        <module>OperatorAndFunctionOfEnumTestModel</module>
         <module>OperatorsModel</module>
+        <module>PagingModel</module>
         <module>PrimitivesModel</module>
         <module>QueryModel</module>
         <module>SalesModels</module>
 <!--        <module>SampleModels</module> -->
         <module>StaticQueryDaoTestModel</module>
+        <module>SpecialCasesModel</module>
         <module>TaskModel</module>
         <module>TernaryTestModel</module>
         <module>TestDefaultAndRequiredFieldsOnMappedTOModel</module>
@@ -52,7 +52,6 @@
         <module>TransferConstructorModel</module>
         <module>RecursiveCompositionModel</module>
         <module>RelationCreateWithAttachmentTestModel</module>
-        <module>PagingModel</module>
     </modules>
 
     <build>
